pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.60.0",
    "fastmcp>=2.10.6",
    "groq>=0.30.0",
    "mcp>=1.12.2",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
]
